[math] depress taichi import logging by forcing using `import_taichi()` utility, move taichi random functions into another file
