+++ brainpy/math/random.py
@@ -70,11 +70,4 @@ from brainpy._src.math.random import (
   rand_like as rand_like,
   randint_like as randint_like,
   randn_like as randn_like,
-
-  taichi_lcg_rand as taichi_lcg_rand, 
-  taichi_uniform_int_distribution as taichi_uniform_int_distribution, 
-  taichi_uniform_real_distribution as taichi_uniform_real_distribution, 
-  taichi_normal_distribution as taichi_normal_distribution,
-  taichi_lfsr88 as taichi_lfsr88,
-
 )
